Add files via upload
This is the week 2 project of the internship- an e-commercce healthcare website which provides health care products such as elderly care, healthcare devices,personal care, food and drinks, mother and baby care,fitness supplements, skin care and yoga. It also provides a BMI calculator which is used to track your body mass index value, a quiz on general health and a random fact generator on health.