Add files via upload
This is my week 3 project at MotionCut- A to-do list webpage which has the functionalities to add a new task, remove a task, check and uncheck a task. It also stores the tasks on a local storage, so that they are retained even if the user refreshes or closes the browser. 
Languages used: 
HTML: For basic web design
CSS:For styling the web page as required
JavaScript: To add the required functionalities