Add files via upload
This is week 3 project of my internship at MotionCut. It is a to-do list application which offers the user to add,delete, check and uncheck tasks. This offers the user to keep a track of the tasks to be completed by the user. Furthermore, it also provides a local storage to store the tasks so that the tasks are retained when the user refreshes or closes the browser. 
The languages used are:
HTML: For basic web page design
CSS: For styling the web page
JavaScript: To add the functionalities of add,deleting and local storage