Add files via upload
This is my fourth week project of my internship at MotionCut. This is a quiz platform- Bookish Trivia which is a quiz platform on literary novels under different categories such as Fiction, Non-Fiction, Thriller, Mystery and Fantasy. The platform is very user friendly and has features such as score,leaderboard and also displaying the correct answer if the chosen is wrong immediately. 
Languages used:
HTML: HTML has been used for creating the basic structure of the website.
CSS: CSS has been used to style the website as per the requirements to make it user friendly and seamless.
JavaScript: JavaScript has been used to add all the required functionalities to the website such as score calculation, leaderboard and marking the correct and wrong answer.